dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Document-createElementNS.html

<!DOCTYPE html>
<title>Document.createElementNS</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var invalidNames = [
    "",
    "1foo",
    "\u0300foo",
    "}foo",
    "f}oo",
    "foo}",
    "\ufffffoo",
    "f\uffffoo",
    "foo\uffff",
    "<foo",
    "foo>",
    "<foo>",
    "f<oo"
  ],
  invalidNSQNameCombinations = [
    ["", ":foo"],
    ["", "foo:"],
    ["", "foo:foo"],
    ["http://oops/", "xml:foo"],
    ["http://oops/", "xmlns"],
    ["http://oops/", "xmlns:foo"],
    ["http://www.w3.org/2000/xmlns/", "xml:foo"],
    ["http://www.w3.org/2000/xmlns/", "foo:xmlns"]
  ];

  invalidNames.forEach(function(name) {
    test(function() {
      assert_throws("INVALID_CHARACTER_ERR", function() { document.createElementNS("", name) });
    }, "Invalid Name: " + name)
  })

  invalidNSQNameCombinations.forEach(function(a) {
    test(function() {
      assert_throws("NAMESPACE_ERR", function() { document.createElementNS(a[0], a[1]) });
    }, "Invalid namespace " + a[0] + " and QName " + a[1])
  })
})
</script>
